requirements.txt
alembic==1.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv
cryptography